[none] refactor main.rs file
- Refactored main.rs to remove ball generation code
to another file in lib
- Refactored Cargo.toml as required

%SOFTWARE
