+++ lib/Cargo.toml
@@ -8,3 +8,4 @@ authors = ["Sen"]
 [dependencies]
 rand="0.3.14"
 rayon = "1.10.0"
+serde_json = "1.0.125"
